Feature/HD-25

Approved-by: Mentor Intern <mentor@novahub.vn>

+++ src/main/java/vn/novahub/helpdesk/repository/LevelRepository.java
@@ -10,7 +10,7 @@ import javax.transaction.Transactional;
 @Transactional
 public interface LevelRepository extends CrudRepository<Level, Long> {
 
-    boolean existsByValueAndAccountIdAndSkillId(int value, long accountId, long skillId);
+    boolean existsByValueAndAccountIdAndSkillId(long value, long accountId, long skillId);
 
     Level getByAccountIdAndSkillId(long accountId, long skillId);
